+++ src/components/pageLayout/PageLayoutWithTabs/index.tsx
@@ -174,9 +174,9 @@ const PageLayout: React.FC<IPageLayoutProps> = ({ icon, title, rightArea, introI
                 </div>
 
                 <div className={'page-header-right-area'} style={{ display: sessionStorage.getItem('menuHide') === '1' ? 'none' : undefined }}>
-                  {introIcon}
+                  <span className='page-layout-intro-container'>{introIcon}</span>
                   {docFn && (
-                    <a onClick={() => docFn()} style={{ marginRight: 16 }}>
+                    <a onClick={() => docFn()} className='page-layout-doc-container'>
                       {t('docs')}
                     </a>
                   )}